# Conflicts:
#	_LowLevelCode/cpp/combine_sqw/combine_sqw.cpp
#	_LowLevelCode/cpp/combine_sqw/exchange_buffer.cpp
#	_test/test_sqw/experiment_sqw.mat
#	horace_core/configuration/@hor_config/hor_config.m
#	horace_core/sqw/@DnDBase/DnDBase.m
#	horace_core/sqw/@data_sqw_dnd/private/make_sqw_data_from_proj_.m
#	horace_core/sqw/@sqw/private/cut_single.m
#	horace_core/sqw/@sqw/sqw.m

+++ _test/common_functions/FakeFAccess.m
@@ -9,7 +9,7 @@ end
 methods
 
     function obj = FakeFAccess(data)
-        obj.fake_data = data;
+        obj.fake_data = single(data);
         obj.npixels_ = size(data, 2);
     end
     %
@@ -21,7 +21,7 @@ methods
             pix_hi = size(obj.fake_data, 2);
         end
         try
-            data = obj.fake_data(:, pix_lo:pix_hi);
+            data = double(obj.fake_data(:, pix_lo:pix_hi));
         catch ME
             switch ME.identifier
             case 'MATLAB:badsubscript'
@@ -34,7 +34,7 @@ methods
     %
     function data = get_pix_at_indices(obj, indices)
         try
-            data = obj.fake_data(:, indices);
+            data = double(obj.fake_data(:, indices));
         catch ME
             switch ME.identifier
             case 'MATLAB:badsubscript'
@@ -46,7 +46,7 @@ methods
     end
     %
     function pix_range = get_pix_range(obj)
-        pix_range = [min(obj.fake_data(1:4,:),[],2),max(obj.fake_data(1:4,:),[],2)]';
+        pix_range = double([min(obj.fake_data(1:4,:),[],2),max(obj.fake_data(1:4,:),[],2)]');
     end
     %
     function new_obj = upgrade_file_format(~)
